[Refactor]: Divided submodules further to outline future architecture.
