+++ src/filetags/cli/cli.py
@@ -5,10 +5,10 @@ import sys
 
 import colorama
 
-from filetags.cli.parser import CliOptions
+from filetags.cli.parser import CliOptions, extract_tags_from_argument
 from filetags.completion import SimpleCompleter
 from filetags.consts import BETWEEN_TAG_SEPARATOR, IS_WINDOWS
-from filetags.Tags import (_get_tag_visual,
+from filetags.tags import (get_tag_visual,
                            check_for_possible_shortcuts_in_entered_tags,
                            print_tag_shortcut_with_numbers)
 
@@ -95,7 +95,7 @@ def ask_for_tags(
         + colorama.Style.RESET_ALL
     )
     print("                     ")
-    print(_get_tag_visual(tags_for_visual))
+    print(get_tag_visual(tags_for_visual))
     print("                     ")
 
     if len(upto9_tags_for_shortcuts) > 0:
