+++ src/filetags/cli/parser/__init__.py
@@ -1,2 +1,3 @@
 from .options import *
 from .processing import *
+from .validate_options import *
