+++ src/filetags/file_operations/__init__.py
@@ -1,0 +1,5 @@
+from .file_operations import *
+from .find_unique_alternative_to_file import *
+from .general import *
+from .links import *
+from .metadata import *
