+++ src/filetags/tags/__init__.py
@@ -1,0 +1,7 @@
+from .repr import *
+from .shortcuts import *
+from .utils import *
+from .VirtualTags import *
+# TODO: Ideally, everything should go through VirtualTags.
+# Currently, at least 'get_tags_from_files_and_subfolders' is used directly
+from .services import *
