+++ src/filetags/tags/services/__init__.py
@@ -1,0 +1,1 @@
+from .TagLocalFilesystem import *
